Create main.yml
This commit is for creating the main.yml file. The main.yml is a file that is used to define the different workflows and jobs that are to be run in a project. It is a part of the CI/CD pipeline and it contains instructions for the automation of the build, test and deployment process of the application.